more trains, and fits to training data (not got testing data yet)

+++ forecast/dataset/loader.py
@@ -31,7 +31,7 @@ def load_monthly_data(paths: list[str]):
 
     time_steps = [str(ts) for ts in time_steps]
 
-    return stacked_data, time_steps
+    return stacked_data.float(), time_steps
 
 
 def load_h5_data(file_path: str, dataset_name: str) -> tuple[np.ndarray, dict]:
